backend/chat/serializers/ChatMessage.py
from rest_framework import serializers
from chat.models.ChatMessage import ChatMessage

class ChatMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ChatMessage
        fields = '__all__'