requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2-binary
Pillow
channels